FIx typescript errors, circular dependencies
