+++ src/Fragment.tsx
@@ -6,11 +6,11 @@ import {
   FragmentParameterMetadata,
 } from 'fig-tree-evaluator'
 import { CustomNodeProps, IconOk, IconCancel } from './_imports'
-import { NodeTypeSelector } from './NodeTypeSelector'
-import { OperatorProps, PropertySelector } from './Operator'
+import { NodeTypeSelector, PropertySelector } from './CommonSelectors'
+import { OperatorProps } from './Operator'
 import { DisplayBar } from './DisplayBar'
 import { getAvailableProperties } from './validator'
-import { Select, SelectOption } from './Select'
+import { Select } from './Select'
 import { useCommon } from './useCommon'
 
 export const Fragment: React.FC<CustomNodeProps<OperatorProps>> = (props) => {
